@@ -12,16 +12,17 @@
 	
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>com.esri.geoevent.solutions.adapter.cot</groupId>
-	<version>10.5.0</version>
+	<version>10.8.1</version>
 	<artifactId>cot-adapter</artifactId>
-	<name>Esri :: AGES :: Solutions :: Adapter :: CoT</name>
+	<name>Esri :: GeoEvent :: Adapter :: CoT</name>
 	<packaging>bundle</packaging>
 	<properties>
 		<contact.address>geoevent@esri.com</contact.address>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.bundle.plugin.version>2.3.6</maven.bundle.plugin.version>
-		<junit.version>4.8.1</junit.version>
-		<jackson.version>1.9.5</jackson.version>
+		<maven.bundle.plugin.version>5.1.1</maven.bundle.plugin.version>
+		<maven.compiler.release>8</maven.compiler.release>
+		<junit.version>4.13.1</junit.version>
+		<jackson.version>2.9.8</jackson.version>
 	</properties>
 	<repositories>
 		<repository>
