@@ -55,7 +55,7 @@ import javax.xml.bind.Marshaller;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.codehaus.jackson.*;
+import com.fasterxml.jackson.core.*;
 
 public class CoTAdapterOutbound extends OutboundAdapterBase {
 
